requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
manim==0.17.3
numpy==1.25.2
Pillow==9.4.0
matplotlib==3.7.3
requests==2.31.0
google-generativeai==0.7.0
pyngrok==6.0.0